Simple Ticket purchasing dialog and intent timeout.

+++ OpenEvent.Web/Models/Transaction/Transaction.cs
@@ -12,7 +12,7 @@ namespace OpenEvent.Web.Models.Transaction
         public DateTime Start { get; set; }
         public DateTime Updated { get; set; }
         public DateTime End { get; set; }
-        public decimal Amount { get; set; }
+        public long Amount { get; set; }
         public bool Paid { get; set; }
         public PaymentStatus Status { get; set; }
 
@@ -29,7 +29,7 @@ namespace OpenEvent.Web.Models.Transaction
         public DateTime Start { get; set; }
         public DateTime Updated { get; set; }
         public DateTime End { get; set; }
-        public decimal Amount { get; set; }
+        public long Amount { get; set; }
         public bool Paid { get; set; }
         public PaymentStatus Status { get; set; }
         public Guid TicketId { get; set; }
